pyproject.toml
[project]
name = "solana-forensics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "flipside>=2.0.8",
    "python-dotenv>=1.1.0",
    "solana>=0.36.6",
    "solders>=0.26.0",
    "uvicorn>=0.34.0",
]
